cleanup: remove some unused classes
